refactor: 환경 변수 상수화 (#2)
