+++ src/constants/env.ts
@@ -1,0 +1,3 @@
+export const SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL;
+export const KAKAO_REST_API_KEY = process.env.KAKAO_REST_API_KEY;
+export const KAKAO_REDIRECT_URI = process.env.KAKAO_REDIRECT_URI;
