+++ src/modules/axios/constants/index.ts
@@ -1,3 +1,2 @@
 export const ACCESS_TOKEN_KEY = 'ACCESS_TOKEN_KEY';
 export const REFRESH_TOKEN_KEY = 'REFRESH_TOKEN_KEY';
-export const SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL;
