+++ src/modules/axios/instances/client.ts
@@ -1,4 +1,4 @@
-import { SERVER_URL } from '@/modules/axios/constants';
+import { SERVER_URL } from '@/constants/env';
 import { clientRequestInterceptor } from '@/modules/axios/interceptors/request/client';
 import { clientResponseErrorInterceptor } from '@/modules/axios/interceptors/response/client';
 import axios from 'axios';
